DesignMenu: Move actions from Tailor
